Allows for six seat aircraft
