+++ NavLog/CardinalTakeOff.json
@@ -36,6 +36,7 @@
       { "wind": 10, "weight": 2200, "multiple": 0.7495, "for50": true },
       { "wind": 20, "weight": 2200, "multiple": 0.4286, "for50": false },
       { "wind": 20, "weight": 2200, "multiple": 0.5291, "for50": true }
-  ]
+  ],
+  "seatCount": 4
 }
